+++ app/src/main/kotlin/fr/gouv/tchap/android/di/HomeserverModule.kt
@@ -1,0 +1,48 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2025. DINUM
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+ * IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
+ * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
+ * OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+package fr.gouv.tchap.android.di
+
+import android.content.Context
+import dev.zacsweers.metro.AppScope
+import dev.zacsweers.metro.BindingContainer
+import dev.zacsweers.metro.ContributesTo
+import dev.zacsweers.metro.Provides
+import fr.gouv.tchap.android.features.enterprise.api.HomeserverConfiguration
+import io.element.android.libraries.di.annotations.ApplicationContext
+import io.element.android.x.R
+
+@BindingContainer
+@ContributesTo(AppScope::class)
+object HomeserverModule {
+    @Provides
+    fun provideHomeserverConfiguration(
+        @ApplicationContext context: Context,
+    ): HomeserverConfiguration {
+        val homeserverList = context.resources.getStringArray(R.array.default_homeservers).toList()
+        return HomeserverConfiguration(
+            defaultHomeserverList = homeserverList
+        )
+    }
+}
