+++ features/login/impl/src/main/kotlin/fr/gouv/tchap/android/features/login/impl/screens/sidentlogin/SidentLoginStateProvider.kt
@@ -22,30 +22,41 @@
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package io.element.android.features.login.impl.screens.sidentlogin
+package fr.gouv.tchap.android.features.login.impl.screens.sidentlogin
 
-import android.os.Parcelable
+import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import io.element.android.features.login.impl.accountprovider.AccountProvider
+import io.element.android.features.login.impl.accountprovider.anAccountProvider
 import io.element.android.features.login.impl.login.LoginMode
 import io.element.android.libraries.architecture.AsyncData
-import kotlinx.parcelize.Parcelize
 
-data class SidentLoginState(
-    val applicationName: String,
-    val accountProvider: AccountProvider,
-    val formState: LoginFormState,
-    val loginMode: AsyncData<LoginMode>,
-    val eventSink: (SidentLoginEvents) -> Unit
-) {
-    val submitEnabled: Boolean
-        get() = accountProvider.url.isNotEmpty() && (loginMode is AsyncData.Uninitialized || loginMode is AsyncData.Loading)
+open class SidentLoginStateProvider : PreviewParameterProvider<SidentLoginState> {
+    override val values: Sequence<SidentLoginState>
+        get() = sequenceOf(
+            aSidentLoginState(),
+            // Loading
+            aSidentLoginState(loginMode = AsyncData.Loading()),
+            // Error
+            aSidentLoginState(loginMode = AsyncData.Failure(Exception("An error occurred"))),
+        )
 }
 
-@Parcelize
-data class LoginFormState(
-    val login: String,
-) : Parcelable {
-    companion object {
-        val Default = LoginFormState("")
-    }
-}
+fun aSidentLoginState(
+    applicationName: String = "Tchap",
+    accountProvider: AccountProvider = anAccountProvider(),
+    formState: LoginFormState = LoginFormState.Default,
+    loginMode: AsyncData<LoginMode> = AsyncData.Uninitialized,
+    eventSink: (SidentLoginEvents) -> Unit = {},
+) = SidentLoginState(
+    applicationName = applicationName,
+    accountProvider = accountProvider,
+    formState = formState,
+    loginMode = loginMode,
+    eventSink = eventSink,
+)
+
+fun aLoginFormState(
+    login: String = "",
+) = LoginFormState(
+    login = login,
+)
