+++ features/login/impl/src/test/kotlin/fr/gouv/tchap/android/features/login/impl/screens/sidentlogin/SidentLoginPresenterTest.kt
@@ -22,7 +22,7 @@
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package io.element.android.features.login.impl.screens.sidentlogin
+package fr.gouv.tchap.android.features.login.impl.screens.sidentlogin
 
 import com.google.common.truth.Truth.assertThat
 import io.element.android.appconfig.AuthenticationConfig
